+++ awot/dropsondes/__init__.py
@@ -11,7 +11,5 @@ Probe Subpackage (:mod:'awot)
 
 """
 from __future__ import absolute_import
-from .thermocalcs import ThermoCalcs
-from .shearcalcs import ShearCalcs
-from .skew import SkewXTick
+
