+++ awot/util/__init__.py
@@ -13,5 +13,7 @@ from .matcher import TrackMatch
 from .convert import (pyart_radar_to_awot, to_awot_flight)
 from .helper import (time_subset_awot_dict, add_dict_to_awot,
                      add_dict_to_awot_fields)
+from .thermocalcs import ThermoCalcs
+from .shearcalcs import ShearCalcs 
 
 __all__ = [s for s in dir() if not s.startswith('_')]
