+++ awot/util/shearcalcs.py
